+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Changelog
 =========
 
 
+- Fix previous fix for case where schema results could have ``data`` or ``errors`` with schema classes as keys.
 - Ensure that ``Select('key', <iteratee>)`` doesn't call ``<iteratee>`` if ``'key'`` was not found in the source data.